GRF-371: Update models
