+++ src/Akeneo/Category/back/Application/Storage/ProcessCategoryUpdateMock.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\Category\Application\Storage;
 
 use Akeneo\Category\Api\Command\UserIntents\UserIntent;
-use Akeneo\Category\Domain\Model\Category;
+use Akeneo\Category\Domain\Model\Enrichment\Category;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
