+++ src/Akeneo/Category/back/Application/Storage/Save/Saver/CategoryBaseSaver.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\Category\Application\Storage\Save\Saver;
 
 use Akeneo\Category\Application\Storage\Save\Query\UpsertCategoryBase;
-use Akeneo\Category\Domain\Model\Category;
+use Akeneo\Category\Domain\Model\Enrichment\Category;
 
 /**
  * This class is used to call the save query for base data of the category (data stored in category table).
