+++ src/Akeneo/Category/back/Domain/Model/Enrichment/Category.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Category\Domain\Model;
+namespace Akeneo\Category\Domain\Model\Enrichment;
 
 use Akeneo\Category\Domain\ValueObject\CategoryId;
 use Akeneo\Category\Domain\ValueObject\Code;
