+++ src/Akeneo/Category/back/tests/Specification/Application/Applier/SetRichTextApplierSpec.php
@@ -7,7 +7,7 @@ namespace Specification\Akeneo\Category\Application\Applier;
 use Akeneo\Category\Api\Command\UserIntents\SetRichText;
 use Akeneo\Category\Api\Command\UserIntents\SetText;
 use Akeneo\Category\Application\Applier\SetRichTextApplier;
-use Akeneo\Category\Domain\Model\Category;
+use Akeneo\Category\Domain\Model\Enrichment\Category;
 use Akeneo\Category\Domain\ValueObject\CategoryId;
 use Akeneo\Category\Domain\ValueObject\Code;
 use Akeneo\Category\Domain\ValueObject\LabelCollection;
