+++ src/Akeneo/Category/back/tests/Specification/Application/Applier/SetTextApplierSpec.php
@@ -7,9 +7,8 @@ namespace Specification\Akeneo\Category\Application\Applier;
 use Akeneo\Category\Api\Command\UserIntents\SetImage;
 use Akeneo\Category\Api\Command\UserIntents\SetText;
 use Akeneo\Category\Application\Applier\SetTextApplier;
-use Akeneo\Category\Application\Applier\SetTextAreaApplier;
 use Akeneo\Category\Application\Applier\UserIntentApplier;
-use Akeneo\Category\Domain\Model\Category;
+use Akeneo\Category\Domain\Model\Enrichment\Category;
 use Akeneo\Category\Domain\ValueObject\CategoryId;
 use Akeneo\Category\Domain\ValueObject\Code;
 use Akeneo\Category\Domain\ValueObject\LabelCollection;
