+++ src/Akeneo/Category/back/tests/Specification/Domain/Model/Enrichment/TemplateSpec.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Specification\Akeneo\Category\Domain\Model;
+namespace Specification\Akeneo\Category\Domain\Model\Enrichment;
 
 use Akeneo\Category\Domain\Model\Attribute\AttributeImage;
 use Akeneo\Category\Domain\Model\Attribute\AttributeRichText;
 use Akeneo\Category\Domain\Model\Attribute\AttributeText;
 use Akeneo\Category\Domain\Model\Attribute\AttributeTextArea;
-use Akeneo\Category\Domain\Model\Template;
+use Akeneo\Category\Domain\Model\Enrichment\Template;
 use Akeneo\Category\Domain\ValueObject\Attribute\AttributeAdditionalProperties;
 use Akeneo\Category\Domain\ValueObject\Attribute\AttributeCode;
 use Akeneo\Category\Domain\ValueObject\Attribute\AttributeCollection;
