+++ src/Akeneo/Category/back/tests/Specification/Infrastructure/Builder/TemplateBuilderSpec.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Specification\Akeneo\Category\Infrastructure\Builder;
 
-use Akeneo\Category\Domain\Model\Category;
+use Akeneo\Category\Domain\Model\Enrichment\Category;
 use Akeneo\Category\Domain\Query\GetCategoryInterface;
 use Akeneo\Category\Domain\ValueObject\Attribute\AttributeType;
 use Akeneo\Category\Domain\ValueObject\CategoryId;
